backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clair-backend"
version = "0.1.0"
description = "Clair — AI-powered note organizer backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "supabase>=2.10",
    "celery[redis]>=5.4",
    "anthropic>=0.40",
    "openai>=1.57",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.6",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
